Fix copilot-review-findings
